requirements.txt
Flask
python-dotenv
mysql-connector-python
Faker
datetime
plotly
pandas